implemented suggested pmier changes
